:recycle: Separated node reference
